pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Gendiff — a difference calculator"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "MetaSveta", email = "mlasouls@gmail.com"},
]

dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
